Outline Of Application Class
The class contains a constructor that will be used to set up any application variables (like the view and logic) and a loop method that will contain the main game loop. This loop function returns an int that represents the fail/success state of the loop that is used by main() to represent the fail/success state of the entire program.